Fix(User) : Memperbaiki hak akses pada halaman User Management.

+++ app/page/user/hapus.php
@@ -1,6 +1,13 @@
 <?php
+if ($currentUser['level'] != 1) {
+    echo "<script>
+       window.location = 'index.php?alert=err2';
+    </script>";
+    exit;
+}
 $pdo = Koneksi::connect();
 $crudUser = user::getInstance($pdo);
+
 $id_user = $_GET['id'];
 
 if ($crudUser->delete($id_user) == true) {
